leo-mar/requirements.txt
edlib==1.3.4
panda
spacy==2.0.18
Flask
gunicorn==20.0.4
